Improve UX: better details, add search for not-tree.
There is a text search, when listing straight.
Search was not yet implemented for tree listing, it's more complicated because branches have to collapse if they have no leaves or branches, and un-collapse if they gain them.

Details are their own page, instead of showing under the list.
This way, it's readable when selected.